% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newGeom.R
\docType{data}
\name{StatCategory}
\alias{StatCategory}
\title{Tile Co-Incidence Matrix}
\format{An object of class \code{StatCategory} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatCategory
}
\arguments{
\item{mapping}{Set of aesthetic mappings created
by aes or aes_. If specified and inherit.aes = TRUE
(the default), is combined with the default mapping at
the top level of the plot. You only need to supply
mapping if there isn't a mapping defined for the plot.}

\item{data}{data frame. If specified, overrides the default
data frame defined at the top level of the plot.}

\item{stat}{The statistical transformation to use on
the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string,
or the result of a call to a position adjustment function.}

\item{na.rm}{f FALSE (the default), removes missing values
with a warning. If TRUE silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in
the legends? NA, the default, includes if any aesthetics are
mapped. FALSE never includes, and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics,
rather than combining with them. This is most useful for helper
functions that define both data and aesthetics and shouldn't inherit
behaviour from the default plot specification}
}
\description{
Input a dataset with at least two categorical variables and the Geom 
results in coincidence square only if a certain pair of values of the two 
variables co-occur.
}
\details{
The GeomTile class has been extended in the following code. The class accepts 
x and y centering values (as center coordinates of the rectangle), aling with 
width and height aesthetics (as dimensions of each rectangle).
}
\examples{
ggplot(mpg, aes(mpg$manufacturer, mpg$class, mpg$hwy)) + 
geom_bloc(aes(height=1,weight=1))


}
\author{
Indraneil Paul, \email{indraneil.paul@research.iiit.ac.in}
}
\keyword{datasets}

